Refactor train pipeline into backend modules
