+++ dvc.yaml
@@ -34,6 +34,9 @@ stages:
       - evaluation.baseline_weights_path
     deps:
       - yolov8_training/train_pipeline.py
+      - yolov8_training/backends/__init__.py
+      - yolov8_training/backends/yolo_backend.py
+      - yolov8_training/backends/rfdetr_backend.py
       - yolov8_training/utils/evaluate.py
       - yolov8_training/utils/data_utils.py
       - yolov8_training/utils/rfdetr_adapter.py
