+++ yolov8_training/backends/__init__.py
@@ -1,0 +1,6 @@
+"""Training backends for the pipeline.
+
+This package contains backend-specific training logic (e.g. YOLO, RF-DETR),
+so that ``train_pipeline.py`` can remain mostly a CLI + stage dispatcher.
+"""
+
